Refactor alarms
