+++ ics/alarm/__init__.py
@@ -1,0 +1,2 @@
+from ics.alarm.audio import AudioAlarm
+from ics.alarm.display import DisplayAlarm
